style: 도메인 디렉토리 이동
