+++ src/main/java/subway/dao/station/StationDao.java
@@ -6,7 +6,7 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import subway.domain.Station;
+import subway.domain.subway.Station;
 import subway.entity.StationEntity;
 
 import javax.sql.DataSource;
