+++ src/main/java/subway/domain/subway/Line.java
@@ -1,0 +1,36 @@
+package subway.domain.subway;
+
+import subway.domain.common.Color;
+import subway.domain.common.LineNumber;
+import subway.domain.common.Name;
+
+public class Line {
+
+    private final Sections sections;
+    private final LineNumber lineNumber;
+    private final Name name;
+    private final Color color;
+
+    public Line(final Sections sections, final Long lineNumber, final String name, final String color) {
+        this.sections = sections;
+        this.lineNumber = new LineNumber(lineNumber);
+        this.name = new Name(name);
+        this.color = new Color(color);
+    }
+
+    public Long getLineNumber() {
+        return lineNumber.getLineNumber();
+    }
+
+    public void addSection(final Section section) {
+        this.sections.addSection(section);
+    }
+
+    public String getName() {
+        return name.getName();
+    }
+
+    public String getColor() {
+        return color.getColor();
+    }
+}
