+++ src/main/java/subway/domain/subway/Station.java
@@ -1,17 +1,19 @@
-package subway.domain;
+package subway.domain.subway;
+
+import subway.domain.common.Name;
 
 import java.util.Objects;
 
 public class Station {
 
-    private final String name;
+    private final Name name;
 
     public Station(final String name) {
-        this.name = name;
+        this.name = new Name(name);
     }
 
     public String getName() {
-        return name;
+        return name.getName();
     }
 
     @Override
