+++ src/main/java/subway/exception/DistanceForkedException.java
@@ -1,8 +1,8 @@
 package subway.exception;
 
-public class DistanceInvalidException extends RuntimeException {
+public class DistanceForkedException extends RuntimeException {
 
-    public DistanceInvalidException() {
+    public DistanceForkedException() {
         super("갈랫길이 생기면 안됩니다. 길이를 다시 입력해주세요.");
     }
 }
