+++ src/main/java/subway/repository/LineRepository.java
@@ -4,9 +4,9 @@ import org.springframework.stereotype.Repository;
 import subway.dao.line.LineDao;
 import subway.dao.section.SectionDao;
 import subway.dao.station.StationDao;
-import subway.domain.Line;
-import subway.domain.Sections;
-import subway.domain.Station;
+import subway.domain.subway.Line;
+import subway.domain.subway.Sections;
+import subway.domain.subway.Station;
 import subway.entity.LineEntity;
 import subway.entity.SectionEntity;
 import subway.entity.StationEntity;
