+++ src/main/java/subway/repository/SectionRepository.java
@@ -4,9 +4,9 @@ import org.springframework.stereotype.Repository;
 import subway.dao.line.LineDao;
 import subway.dao.section.SectionDao;
 import subway.dao.station.StationDao;
-import subway.domain.Section;
-import subway.domain.Sections;
-import subway.domain.Station;
+import subway.domain.subway.Section;
+import subway.domain.subway.Sections;
+import subway.domain.subway.Station;
 import subway.entity.LineEntity;
 import subway.entity.SectionEntity;
 import subway.entity.StationEntity;
